feat(slider): ticks feat as a separate component
Closes #4594
