+++ src/app/slider/slider.sample.html
@@ -9,9 +9,9 @@
                 <igx-slider #slider [value]="80" [step]="5" [lowerBound]="0" [tickLabelsOrientation]="labelOrientaion" [primaryTicks]="5" [secondaryTicks]="2">
                 </igx-slider>
             </article>
-            <!-- <article style="padding-top: 50px">
+            <article style="padding-top: 50px">
                 <button igxButton (click)="changeLabelOrientation()">change label orientaiton</button>
-            </article> -->
+            </article>
 
 
             <!-- <article class="sample-column">
